# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ursol CAST API - ERP System",
        "description": "API completa de ERP para empresas costarricenses. Sistema integral de gestión empresarial con soporte para contabilidad, facturación electrónica, inventario, compras, ventas, recursos humanos y más.",
        "contact": {
            "name": "Sistemas Ursol",
            "email": "soporte@ursol.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de Desarrollo Local"
        },
        {
            "url": "https://api.ursol.com",
            "description": "Servidor de Producción"
        }
    ],
    "paths": {
...
        "/api/sucursales": {
            "get": {
                "tags": [
                    "Sucursales"
                ],
                "summary": "Listar sucursales",
                "description": "Obtiene listado paginado de sucursales filtradas por empresa",
                "operationId": "a201ddde2bc11ea5d3a36d8c5b399074",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Registros por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "empresa_id",
                        "in": "query",
                        "description": "Filtrar por empresa",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "activos",
                        "in": "query",
                        "description": "Solo sucursales activas",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Sucursal"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sucursales"
                ],
                "summary": "Crear sucursal",
                "description": "Crea una nueva sucursal. Si es principal, desmarca las demás",
                "operationId": "ba55df2d172c227845f0c4a3830328e1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "empresa_id",
                                    "nombre"
                                ],
                                "properties": {
                                    "empresa_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Sucursal Centro"
                                    },
                                    "codigo": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "provincia": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "canton": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "distrito": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "es_principal": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "activo": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sucursal creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Sucursal"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/sucursales/{id}": {
            "get": {
                "tags": [
                    "Sucursales"
                ],
                "summary": "Obtener sucursal",
                "description": "Detalles de una sucursal con almacenes y cajas",
                "operationId": "e565921e543f7a40c8de275bfccc803f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucursal encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Sucursal"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Sucursales"
                ],
                "summary": "Actualizar sucursal",
                "description": "Actualiza datos de sucursal. Si se marca como principal, desmarca las demás",
                "operationId": "3b60ed1087b05432da52eae1c0c06198",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "es_principal": {
                                        "type": "boolean"
                                    },
                                    "activo": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizada exitosamente"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sucursales"
                ],
                "summary": "Eliminar sucursal",
                "description": "Soft delete. No permite eliminar la sucursal principal",
                "operationId": "b113ba73735a9ee01ac00b94400590fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminada exitosamente"
                    },
                    "404": {
                        "description": "No encontrada"
                    },
                    "422": {
                        "description": "No se puede eliminar la principal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Empresa": {
                "title": "Empresa",
                "description": "Empresa del sistema (tenant)",
                "required": [
                    "id",
                    "nombre",
                    "identificacion"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Sistemas Ursol S.A."
                    },
                    "nombre_comercial": {
                        "type": "string",
                        "example": "Ursol",
                        "nullable": true
                    },
                    "identificacion": {
                        "type": "string",
                        "example": "3-101-123456"
                    },
                    "tipo_identificacion": {
                        "type": "string",
                        "enum": [
                            "CEDULA_FISICA",
                            "CEDULA_JURIDICA",
                            "DIMEX",
                            "NITE"
                        ],
                        "example": "CEDULA_JURIDICA"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "2222-3333",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "info@ursol.com",
                        "nullable": true
                    },
                    "activo": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "Sucursal": {
                "title": "Sucursal",
                "description": "Sucursal de una empresa",
                "required": [
                    "id",
                    "empresa_id",
                    "nombre"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "empresa_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Sucursal Centro"
                    },
                    "codigo": {
                        "type": "string",
                        "example": "SUC-001",
                        "nullable": true
                    },
                    "telefono": {
                        "type": "string",
                        "example": "2222-3333",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "centro@empresa.com",
                        "nullable": true
                    },
                    "direccion": {
                        "type": "string",
                        "example": "San José, Centro, Avenida Central",
                        "nullable": true
                    },
                    "provincia": {
                        "type": "string",
                        "example": "San José",
                        "nullable": true
                    },
                    "canton": {
                        "type": "string",
                        "example": "San José",
                        "nullable": true
                    },
                    "distrito": {
                        "type": "string",
                        "example": "Carmen",
                        "nullable": true
                    },
                    "es_principal": {
                        "type": "boolean",
                        "example": true
                    },
                    "activo": {
                        "type": "boolean",
                        "example": true
                    },
                    "eliminado": {
                        "type": "boolean",
                        "example": false
                    },
                    "creado_en": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "actualizado_en": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "empresa": {
                        "$ref": "#/components/schemas/Empresa"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum Bearer Token Authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}